fix mtp clean code warning

+++ frameworks/services/media_mtp/src/payload_data/get_device_prop_value_data.cpp
@@ -59,7 +59,7 @@ uint32_t GetDevicePropValueData::CalculateSize()
 
     int res = Maker(tmpuse);
     if (res != MTP_SUCCESS) {
-        return res;
+        return static_cast<uint32_t>(res);
     }
 
     uint32_t size = tmpuse.size();
